Extend connection configuration
Add ‘recovery_attempts’ to the connection configuration with a default
of 10 attempts.

In the past we observed some applications that did not re-connection
to RabbitMQ properly. By providing this option and giving it a
reasonable default, we want to make sure that an application crashes
and is restarted by Kubernetes every time runs into such a stale
state.