Modify worker thread behaviour
The session reconnection logic of Bunny seems to happen in the
reader_loop which is kind of strange and unexpected. The reader_loop
thread itself won’t propagate exceptions to the main thread.  Hence,
when we raise an exception in the recovery_attempts_exhausted Proc, it
won’t terminate the app unless we modify the thread behaviour
beforehand.